backend/requirements.txt
Django==4.2.20
django-constance==4.3.2
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
pillow==10.4.0
django-environ
django-model-utils